requirements.txt
fastapi>=0.115,<0.120
uvicorn[standard]>=0.30,<0.40
websockets>=14.0,<16.0
python-multipart>=0.0.12
python-dotenv>=1.0
jinja2>=3.1
aiofiles>=24.1
httpx>=0.27
ollama>=0.4
speechmatics-python>=3.0
numpy>=1.26
qrcode[pil]>=7.4
